Setup Continuous Integration (#3)
* Gitlab CI -> Github Actions

* install ninja

* disable windows build and test for now

* apt-get -> sudo apt-get

* python3.6 -> python on ubuntu-latest

* don't use ubuntu:18:04 docker image

* explicit artifact path

* setup pythonversion

* remove getspaindata

* Add CI status badge

* change to Ubuntu 18.04 for CPP jobs

* don't build as root in docker?

* don't build in conanio/gcc72 at all

* maybe no parallel build?

* test on 18.04

* sudo

* artifact paths

* weird permission issue with unit tests

* remove js coverage

* remove gitlab ci

* Update README.md

* switch codecov action v1 -> v2

* fix coding guidelines link in PR template

* Update .github/pull_request_template.md


* Update .github/pull_request_template.md


* Update .github/pull_request_template.md


* Update .github/pull_request_template.md


* Update .github/pull_request_template.md


* remove tixi from CI

* remove reference to undefined testing strategy in PR template

* reenable windows build, install hdf5

* reenable windows build, install hdf5

* actually build with MSVC...

* build step uses cmd instead of ps

* fix compiler warning C4544

doubly defined default template argument

* explicit conversion for happy msvc

* make msvc happy

* change title of README

* add disable_autolinking to public interface of EPI_IO

this is needed to disable boost autolinking when linking the unit tests against epidemiology_io

* checkout code in win tests

* artifact whole build directory for linux test

+++ cpp/epidemiology/secir/analyze_result.h
@@ -127,7 +127,7 @@ std::vector<Simulation> ensemble_params_percentile(const std::vector<std::vector
     std::vector<double> single_element_ensemble(num_runs);
 
     // lamda function that calculates the percentile of a single paramter
-    std::vector<Simulation> percentile(num_nodes, Simulation(num_groups));
+    std::vector<Simulation> percentile(num_nodes, Simulation((int)num_groups));
     auto param_percentil = [&ensemble_params, p, num_runs, &percentile](auto n, auto get_param) mutable {
         std::vector<double> single_element(num_runs);
         for (size_t run = 0; run < num_runs; run++) {
